refactor version into its own class
